Add files via upload
In this work, I have taken Fashion-MNIST dataset for building CNN. In CNN, I used sequential, dense and flatten layers. Basically, a CNN can be instantiated as a Sequential model because each layer has exactly one input and output and is stacked together to form the entire network. Dense Layer is simple layer of neurons in which each neuron receives input from all the neurons of previous layer, thus called as dense. Dense Layer is used to classify image based on output from convolutional layers. Flatten is used to flatten the input. The Flatten layer converts the 28x28x50 output of the convolutional layer into a single one-dimensional vector, that can be used as input for a dense layer. And also I compared the accuracy of the CNN model for different optimizers such as Adam, RMS, Adagrad, SGD and SGDM. Out of them Adam gave high accuracy of 91%